requirements.txt
Django==5.0.7
requests==2.32.3
djangorestframework==3.15.2
markdown==3.6
django-filter==24.2
drf-yasg==1.21.7
django-environ==0.11.2